Rename more components to kebab case
