+++ src/components/configuration-parameters/configuration-parameters.tsx
@@ -1,6 +1,6 @@
 import TextField from '@material-ui/core/TextField';
 import * as React from 'react';
-import * as styles from './configurationParameters.styles';
+import * as styles from './configuration-Parameters.styles';
 
 export const ConfigurationParameters: React.FunctionComponent = () => (
   <div className={styles.configurationParameters}>
