+++ src/components/configuration-parameters/index.ts
@@ -1,0 +1,1 @@
+export * from './configuration-parameters';
