+++ src/components/configuration/configuration.tsx
@@ -1,7 +1,7 @@
 import Tab from '@material-ui/core/Tab';
 import Tabs from '@material-ui/core/Tabs';
 import * as React from 'react';
-import { ConfigurationParameters } from '../configurationParameters';
+import { ConfigurationParameters } from '../configuration-parameters';
 import { SettingsState } from '../settingsState';
 import { Severity } from '../severity';
 import { TypeUse } from '../typeUse';
