--- src/components/configurationParameters/index.ts
@@ -1,1 +1,0 @@
-export * from './configurationParameters';
