+++ src/components/search-bar/search-bar.styles.ts
