+++ src/components/search-bar/search-bar.tsx
@@ -1,7 +1,7 @@
 import { Input, Paper } from '@material-ui/core';
 import SearchIcon from '@material-ui/icons/Search';
 import * as React from 'react';
-import * as styles from './searchbar.styles';
+import * as styles from './search-bar.styles';
 
 export const Searchbar: React.StatelessComponent = () => (
   <Paper className={styles.searchbar}>
