--- src/components/searchbar/index.ts
@@ -1,1 +1,0 @@
-export * from './searchbar';
